requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pydantic>=2.0.0
requests>=2.31.0

# Optional integrations
pinecone-client>=2.2.0
notion-client>=2.0.0
